physic world for loop size_t bugfix

+++ triebWerk/src/CPhysicWorld.cpp
@@ -211,7 +211,7 @@ void triebWerk::CPhysicWorld::CheckCollisionEvents()
     {
         ICollider* collider = m_DynamicCollider[i];
 
-        for (size_t j = collider->m_CollisionEvents.size() - 1; j >= 0; --j)
+        for (int j = collider->m_CollisionEvents.size() - 1; j >= 0; --j)
         {
             CCollisionEvent& collEvent = collider->m_CollisionEvents[j];
